Add c++ test

+++ test/test.sh
@@ -5,6 +5,11 @@ set -e
 TEST=$1
 INSTR_FILE=$(find ../test -type f -name $TEST.txt)
 
+if [ -z $INSTR_FILE ]; then
+	echo "No $TEST.txt found"
+	exit 1
+fi
+
 echo "Running test $TEST"
 
 rm -f $TEST.out
